Merge in LFAC/secure_enclave from LF-11545 to integration

* commit 'c835a8a983935bb24bc613f957d849dc847945c3':
  LF-11545: Handle Expected failures of Signature Generation and Verification RSA test vectors

+++ test/common/include/test_common_tv.h
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 /*
- * Copyright 2022-2023 NXP
+ * Copyright 2022-2024 NXP
  */
 
 #ifndef TEST_COMMON_TV_H
@@ -18,20 +18,22 @@
 #define DEFAULT_TV_PKF_PATH "/var/lib/se/persistent/"
 #define DEFAULT_TV_PKF_FPATH (DEFAULT_TV_PKF_PATH DEFAULT_TV_PKF_FNAME)
 
-void generate_key_test_tv(hsm_hdl_t key_store_hdl, FILE *fp, char *line, uint8_t
-			  *tests_passed, uint8_t *tests_failed,
-			  uint8_t *tests_invalid, uint8_t *tests_total);
+void generate_key_test_tv(hsm_hdl_t key_store_hdl, FILE *fp, char *line,
+			  uint16_t *tests_passed, uint16_t *tests_failed,
+			  uint16_t *tests_invalid, uint16_t *tests_total);
 void cipher_test_tv(hsm_hdl_t key_store_hdl, FILE *fp, char *line,
-		    uint8_t *tests_passed, uint8_t *tests_failed,
-		    uint8_t *tests_invalid, uint8_t *tests_total);
-void mac_test_tv(hsm_hdl_t key_store_hdl, FILE *fp, char *line, uint8_t *tests_passed,
-		 uint8_t *tests_failed, uint8_t *tests_invalid, uint8_t *tests_total);
+		    uint16_t *tests_passed, uint16_t *tests_failed,
+		    uint16_t *tests_invalid, uint16_t *tests_total);
+void mac_test_tv(hsm_hdl_t key_store_hdl, FILE *fp, char *line,
+		 uint16_t *tests_passed, uint16_t *tests_failed,
+		 uint16_t *tests_invalid, uint16_t *tests_total);
 void sign_verify_test_tv(hsm_hdl_t key_store_hdl, FILE *fp, char *line,
-			 uint8_t *tests_passed, uint8_t *tests_failed,
-			 uint8_t *tests_invalid, uint8_t *tests_total);
-void hash_test_tv(FILE *fp, char *line, uint8_t *tests_passed, uint8_t *tests_failed,
-		  uint8_t *tests_invalid, uint8_t *tests_total);
+			 uint16_t *tests_passed, uint16_t *tests_failed,
+			 uint16_t *tests_invalid, uint16_t *tests_total);
+void hash_test_tv(FILE *fp, char *line,
+		  uint16_t *tests_passed, uint16_t *tests_failed,
+		  uint16_t *tests_invalid, uint16_t *tests_total);
 void data_storage_test_tv(hsm_hdl_t key_store_hdl, FILE *fp, char *line,
-			  uint8_t *tests_passed, uint8_t *tests_failed,
-			  uint8_t *tests_invalid, uint8_t *tests_total);
+			  uint16_t *tests_passed, uint16_t *tests_failed,
+			  uint16_t *tests_invalid, uint16_t *tests_total);
 #endif
