+++ test/common/test_hash_tv.c
@@ -322,14 +322,15 @@ exit:
 	return test_status;
 }
 
-void hash_test_tv(FILE *fp, char *line, uint8_t *tests_passed, uint8_t *tests_failed,
-		  uint8_t *tests_invalid, uint8_t *tests_total)
+void hash_test_tv(FILE *fp, char *line,
+		  uint16_t *tests_passed, uint16_t *tests_failed,
+		  uint16_t *tests_invalid, uint16_t *tests_total)
 {
 	int8_t test_status = TEST_STATUS_FAILED;
-	static uint8_t thash_passed;
-	static uint8_t thash_failed;
-	static uint8_t thash_invalids;
-	static uint8_t thash_total;
+	static uint16_t thash_passed;
+	static uint16_t thash_failed;
+	static uint16_t thash_invalids;
+	static uint16_t thash_total;
 
 	int len = strlen(line);
 	char *test_id = (char *)malloc(len * sizeof(char));
